Add UnitTest for PointCloudXYZ, PointCloudXYZI, PointCloudXYZRGBA
Add a TryLoad case in Invoke.OnImport function to look for native dll in native folder. Now Pcl.NET can correctly be linked in debug and both managed and debug dlls can be step into

Fix LoadPointCloudXYZI to use correct io_load_pcd_xyzi function

Add UnorganizedPointCloudException and add some checking and throwing

Add some comments

+++ src/Pcl.NET/PointXYZI.cs
@@ -9,6 +9,12 @@ using System.Threading.Tasks;
 
 namespace Pcl.NET
 {
+    /// <summary>
+    /// Represents a three-dimensional point with an associated intensity value
+    /// data processing.
+    /// </summary>
+    /// <remarks>The PointXYZI structure stores the X, Y, and Z coordinates along with an intensity value,
+    /// allowing efficient access to both individual components and the combined vector representation.</remarks>
     [DebuggerDisplay("{V}, {Label}")]
     [StructLayout(LayoutKind.Explicit, Size = 32)]
     public unsafe struct PointXYZI : IEquatable<PointXYZI>
@@ -26,9 +32,9 @@ namespace Pcl.NET
         public Vector3 V;
 
         [FieldOffset(16)]
-        public uint Intensity;
+        public float Intensity;
 
-        public PointXYZI(float x, float y, float z, uint intensity) : this()
+        public PointXYZI(float x, float y, float z, float intensity) : this()
         {
             X = x;
             Y = y;
